Move basic usage instruction into the program
The user of the program shall not require a copy of the README file to
use the program.

Also move copyright and version information into the program source
code.

It does not seem to be a good idea to maintain the same information in
two different files which are both put under version control.

Finally, add support for arbitrarily-sized key files. If the file size
is not an even number of octets, just assume another null octet was
present after the actual end of the file in order to make it even.